+++ ddGetChunk_snippet.php
@@ -1,68 +1,23 @@
 <?php
 /**
  * ddGetChunk
- * @version 2.2.2 (2020-06-22)
+ * @version 2.3 (2021-03-24)
  * 
  * @see README.md
  * 
  * @link https://code.divandesign.biz/modx/ddgetchunk
  * 
- * @copyright 2009–2020 DD Group {@link https://DivanDesign.biz }
+ * @copyright 2009–2021 DD Group {@link https://DivanDesign.biz }
  */
 
-//The snippet must return an empty string even if result is absent
-$snippetResult = '';
+//Include (MODX)EvolutionCMS.libraries.ddTools
+require_once(
+	$modx->getConfig('base_path') .
+	'assets/libs/ddTools/modx.ddtools.class.php'
+);
 
-if (!empty($name)){
-	//Include (MODX)EvolutionCMS.libraries.ddTools
-	require_once(
-		$modx->getConfig('base_path') .
-		'assets/libs/ddTools/modx.ddtools.class.php'
-	);
-	
-	//Для обратной совместимости
-	extract(\ddTools::verifyRenamedParams([
-		'params' => $params,
-		'compliance' => [
-			'escapeResultForJS' => 'escaping'
-		]
-	]));
-	
-	//Получаем чанк
-	$snippetResult = $modx->getTpl($name);
-	
-	//Если переданы дополнительные данные
-	if (!empty($placeholders)){
-		$placeholders = \ddTools::encodedStringToArray($placeholders);
-	}else{
-		$placeholders = [];
-	}
-	
-	//Парсим
-	$snippetResult = \ddTools::parseText([
-		'text' => $snippetResult,
-		'data' => $placeholders,
-		//Удаляем пустые плэйсхолдеры, если нужно
-		'removeEmptyPlaceholders' =>
-			(
-				isset($removeEmptyPlaceholders) &&
-				$removeEmptyPlaceholders == '1'
-			) ?
-			true :
-			false
-	]);
-	
-	//Окончательно парсим
-	$snippetResult = \ddTools::parseSource($snippetResult);
-	
-	//Экранируем сиволы, если нужно
-	if (
-		isset($escapeResultForJS) &&
-		$escapeResultForJS == '1'
-	){
-		$snippetResult = \ddTools::escapeForJS($snippetResult);
-	}
-}
-
-return $snippetResult;
+return \DDTools\Snippet::runSnippet([
+	'name' => 'ddGetChunk',
+	'params' => $params
+]);
 ?>